Baruuuuuuuuuuuuuuuuuuuu

+++ routes/web.php
@@ -47,10 +47,13 @@ Route::middleware('auth')->group(function () {
     // users
     Route::get('/Users', [UsersController::class, 'index'] )->name('Users')->middleware(['auth', 'verified', 'permission:lihat-user']);
     Route::get('/Users/create', [UsersController::class, 'create'])->name('Users.tambah')->middleware(['auth', 'verified', 'permission:tambah-user']);
-    Route::post('/Users/store', [UsersController::class, 'store'])->name('Users.store');
+    Route::post('/Users/store', [UsersController::class, 'store'])->name('Users.store')->middleware(['auth', 'verified', 'permission:lihat-user']);
     Route::get('/Users/edit/{id}', [UsersController::class, 'edit'])->name('Users.edit')->middleware(['auth', 'verified', 'permission:edit-user']);
-    Route::put('/Users/update/{id}', [UsersController::class, 'update'])->name('Users.update');
-    Route::delete('/Users/delete/{id}', [UsersController::class, 'delete'])->name('Users.delete');
+    Route::put('/Users/update/{id}', [UsersController::class, 'update'])->name('Users.update')->middleware(['auth', 'verified', 'permission:lihat-user']);
+    Route::delete('/Users/delete/{id}', [UsersController::class, 'delete'])->name('Users.delete')->middleware(['auth', 'verified', 'permission:lihat-user']);
+    Route::get('/Users/invoice', [UsersController::class, 'invoice'] )->name('Users.invoice');
+
+    //Prodile
     Route::get('/Profile/{id}', [UsersController::class, 'profile'] )->name('Profile');
     Route::put('/Profile/update/{id}', [UsersController::class, 'updateProfile'])->name('Profile.update');
 
@@ -61,6 +64,7 @@ Route::middleware('auth')->group(function () {
     Route::get('/Ajaran/edit/{kode_ajaran}', [AjaranController::class, 'edit'])->name('Ajaran.edit');
     Route::put('/Ajaran/update/{kode_ajaran}', [AjaranController::class, 'update'])->name('Ajaran.update');
     Route::delete('/Ajaran/delete/{kode_ajaran} ', [AjaranController::class, 'delete'])->name('Ajaran.delete');
+    Route::get('/Ajaran/invoice', [AjaranController::class, 'invoice'] )->name('Ajaran.invoice');
 
     // Data Kelas
     Route::get('/Data-kelas', [DataKelasController::class, 'index'] )->name('Data-kelas');
@@ -69,6 +73,7 @@ Route::middleware('auth')->group(function () {
     Route::get('/Data-kelas/edit/{id_kelas}', [DataKelasController::class, 'edit'])->name('Data-kelas.edit');
     Route::put('/Data-kelas/update/{id_kelas}', [DataKelasController::class, 'update'])->name('Data-kelas.update');
     Route::delete('/Data-kelas/delete/{id_kelas}', [DataKelasController::class, 'delete'])->name('Data-kelas.delete');
+    Route::get('/Data-kelas/invoice', [DataKelasController::class, 'invoice'] )->name('Data-kelas.invoice');
 
     // Setting
     Route::get('/Setting', [SettingController::class, 'index'] )->name('Setting')->middleware(['auth', 'verified', 'permission:lihat-setting']);
@@ -82,20 +87,24 @@ Route::middleware('auth')->group(function () {
     Route::get('/Pembayaran', [PembayaranController::class, 'index'])->name('Pembayaran');
     Route::get('/Pembayaran/edit/{id_pembayaran}', [PembayaranController::class, 'edit'])->name('Pembayaran.edit');
     Route::put('/Pembayaran/update/{id_pembayaran}', [PembayaranController::class, 'update'])->name('Pembayaran.update');
+    Route::get('/Pembayaran/invoice', [PembayaranController::class, 'invoice'] )->name('Pembayaran.invoice');
 
-    // History  Pembayaran
+    // Tingkatan
     Route::get('/Tingkatan', [TingkatController::class, 'index'] )->name('Tingkatan');
     Route::get('/Tingkatan/create', [TingkatController::class, 'create'] )->name('Tingkatan.tambah');
     Route::post('/Tingkatan/store', [TingkatController::class, 'store'] )->name('Tingkatan.store');
     Route::get('/Tingkatan/edit/{kode_tingkat}', [TingkatController::class, 'edit'] )->name('Tingkatan.edit');
     Route::put('/Tingkatan/update/{kode_tingkat}', [TingkatController::class, 'update'] )->name('Tingkatan.update');
     Route::delete('/Tingkatan/delete/{kode_tingkat}', [TingkatController::class, 'delete'] )->name('Tingkatan.delete');
+    Route::get('/Tingkatan/invoice', [TingkatController::class, 'invoice'] )->name('Tingkatan.invoice');
 
     // History  Pembayaran
     Route::get('/History', [PembayaranController::class, 'history'] )->name('History');
+    Route::get('/History/invoice', [PembayaranController::class, 'invoice'] )->name('History.invoice');
 
     // Tunggakan
     Route::get('/Tunggakan', [PembayaranController::class, 'tunggakan'] )->name('Tunggakan');
+    Route::get('/Tunggkan/invoice', [PembayaranController::class, 'invoice'] )->name('Tunggakan.invoice');
 
 });
